% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_id_by_shows.R
\name{dub_id_by_shows}
\alias{dub_id_by_shows}
\title{IDs for shows}
\format{Returns a list of \code{dub_id}s.}
\usage{
dub_id_by_shows(shows)
}
\arguments{
\item{shows}{A show or a list of shows for which we'd like to return the \code{dub_id}s.}
}
\value{
vector
}
\description{
Get the list of IDs that are in scope for the selected set of shows.
}
\details{
This would tipically be used to filter for texts or characters of a specific show.
}
\examples{
dub_id_by_shows("Fifth Gear")
dub_id_by_shows(c("Fifth Gear", "Finding Bigfoot"))

}
\seealso{
\code{\link{dub_text}}, \code{\link{dub_characters}}
}
