% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_metadata.R
\name{dub_metadata}
\alias{dub_metadata}
\title{Metadata of the scripts}
\format{Returns a tibble with the metadata.
\itemize{
    \item \code{dub_id}: Unique identifier of the scripts.
    \item \code{production_code}: The production code that was used by the production company.
    \item \code{show}: The name of the TV show.
    \item \code{season}: The season of the TV show for which the translation was requested.
    \item \code{episode}: The episode of the TV show within the season for which the translation was requested.
}}
\usage{
dub_metadata()
}
\description{
Metadata about audiovisual translations. The scripts were translated by Patrik Szigeti.
}
\examples{
dub_metadata()

}
\seealso{
dub_text, dub_characters, dub_shows
}
