% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_shows.R
\name{dub_shows}
\alias{dub_shows}
\title{Shows that appear in the dubbR package}
\format{Returns a tibble with the unique list of shows that appear in the package.
\itemize{
    \item \code{show}: The unique TV shows in the package.
}}
\usage{
dub_shows()
}
\value{
tibble
}
\description{
A unique list of the shows that appear in the package. It can be used to explore the package and filter the contents.
}
\examples{
dub_shows()

}
\seealso{
\code{\link{dub_metadata}}, \code{\link{dub_text}}, \code{\link{dub_characters}}
}
