% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_text.R
\name{dub_text}
\alias{dub_text}
\title{Text of the scripts}
\format{Returns a tibble with the text of the scripts.
\itemize{
    \item \code{dub_id}: Unique identifier of the scripts.
    \item \code{text}: The text from the scripts line by line.
}}
\usage{
dub_text(dub_id)
}
\arguments{
\item{dub_id}{A list of IDs for which the texts should be imported.}
}
\value{
tibble
}
\description{
The text of the audiovisual translations, formatted to be convenient for text analysis.
}
\examples{
dub_text()
dub_text(c(1, 2))
dub_text(dub_id_by_shows("Fifth Gear"))

}
\seealso{
\code{\link{dub_metadata}}, \code{\link{dub_characters}}, \code{\link{dub_shows}}, \code{\link{dub_id_by_shows}}
}
