% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_data.R
\name{translation_info}
\alias{translation_info}
\title{Data wrangling of the scripts}
\usage{
translation_info(file_name)
}
\description{
Creates the tibbles that should be returned for the separate functions
}
