% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_characters.R
\name{dub_characters}
\alias{dub_characters}
\title{Characters of the scripts}
\format{Returns a tibble with the characters from the scripts.
\itemize{
    \item \code{dub_id}: Unique identifier of the scripts.
    \item \code{character}: The characters that appear in the scripts.
}}
\usage{
dub_characters(shows)
}
\arguments{
\item{dub_id}{A list of IDs for which the characters should be imported.}
}
\value{
tibble
}
\description{
The characters from the audiovisual translations. This can be used to create \code{anti_join}s so that the character names wouldn't skew the analysis.
}
\examples{
dub_characters()
dub_characters(c(1, 2))
dub_characters(dub_id_by_shows("Fifth Gear"))

}
\seealso{
\code{\link{dub_metadata}}, \code{\link{dub_text}}, \code{\link{dub_shows}}, \code{\link{dub_id_by_shows}}
}
