% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dub_text.R
\name{dub_text}
\alias{dub_text}
\title{Text of the scripts}
\format{Returns a tibble with the text of the scripts.
\itemize{
    \item \code{dub_id}: Unique identifier of the scripts.
    \item \code{text}: The text from the scripts line by line.
}}
\usage{
dub_text()
}
\description{
The text of the audiovisual translations, formatted to be convenient for text analysis.
}
\examples{
dub_text()

}
\seealso{
dub_metadata, dub_characters, dub_shows
}
